old_projects/book_converter/requirements.txt
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
loguru==0.7.3
lxml==6.0.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
pycparser==2.23
PyMuPDF==1.26.5
PyPDF2==3.0.1
pypdfium2==5.0.0
pytesseract==0.3.13
python-docx==1.2.0
typing_extensions==4.15.0
win32_setctime==1.2.0
